break up layerpane into multiple files, use material slider for opacity
