+++ frontend/package.json
@@ -10,11 +10,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
     "@geoman-io/leaflet-geoman-free": "^2.14.2",
     "@headlessui/react": "latest",
     "@heroicons/react": "^2.0.18",
     "@hookform/resolvers": "^3.6.0",
     "@mapbox/shp-write": "^0.4.3",
+    "@mui/material": "^6.1.4",
     "@nivo/bar": "^0.87.0",
     "@nivo/calendar": "^0.87.0",
     "@nivo/line": "^0.87.0",
