+++ frontend/src/components/maps/LayerPane/index.tsx
@@ -1,0 +1,3 @@
+import LayerPane from './LayerPane';
+
+export default LayerPane;
